removes pycrypto (dead) lib as a dependency and only uses cryptography
