+++ blatann/examples/glucose_peripheral.py
@@ -158,4 +158,4 @@ def main(serial_port):
 
 
 if __name__ == '__main__':
-    main("COM5")
+    main("COM3")
