+++ blatann/gap/smp.py
@@ -1,11 +1,8 @@
 import logging
 import threading
 import enum
-import binascii
-from Crypto.Cipher import AES
-from cryptography.hazmat.backends import default_backend
-from cryptography.hazmat.primitives.asymmetric import ec
 
+from blatann.gap import smp_crypto
 from blatann.gap.bond_db import BondingData
 from blatann.nrf import nrf_types, nrf_events
 from blatann.exceptions import InvalidStateException, InvalidOperationException
@@ -25,9 +22,6 @@ SecurityStatus = nrf_types.BLEGapSecStatus
 # Enum of the different Pairing passkeys to be entered by the user (passcode, out-of-band, etc.)
 AuthenticationKeyType = nrf_types.BLEGapAuthKeyType
 
-# Elliptic Curve used for LE Secure connections
-_lesc_curve = ec.SECP256R1
-
 
 class SecurityLevel(enum.Enum):
     """
@@ -38,80 +32,6 @@ class SecurityLevel(enum.Enum):
     JUST_WORKS = 2
     MITM = 3
     LESC_MITM = 4
-
-
-def ah(key, p_rand):
-    """
-    Function for calculating the ah() hash function described in Bluetooth core specification 4.2 section 3.H.2.2.2.
-
-    This is used for resolving private addresses where a private address
-    is prand[3] || aes-128(irk, prand[3]) % 2^24
-
-    :param key: the IRK to use, in big endian format
-    :param p_rand:
-    :return:
-    """
-    if len(p_rand) != 3:
-        raise ValueError("Prand must be a str or bytes of length 3")
-    cipher = AES.new(key, AES.MODE_ECB)
-
-    # prepend the prand with 0's to fill up a 16-byte block
-    p_rand = chr(0) * 13 + p_rand
-    # Encrypt and return the last 3 bytes
-    return cipher.encrypt(p_rand)[-3:]
-
-
-def pubkey_to_raw(public_key):
-    """
-    Converts from a python public key to the raw (x, y) bytes for the nordic
-    """
-    pk = public_key.public_numbers()
-    x = bytearray.fromhex("{:064x}".format(pk.x))
-    y = bytearray.fromhex("{:064x}".format(pk.y))
-
-    # Nordic requires keys in little-endian, rotate
-    pubkey_raw = x[::-1] + y[::-1]
-    return pubkey_raw
-
-
-def pubkey_from_raw(raw_key):
-    """
-    Converts from raw (x, y) bytes to a public key that can be used for the DH request
-    """
-    key_len = len(raw_key)
-    x_raw = raw_key[:key_len/2]
-    y_raw = raw_key[key_len/2:]
-
-    # Nordic transmits keys in little-endian, convert to big-endian
-    x_raw = x_raw[::-1]
-    y_raw = y_raw[::-1]
-
-    x = int(binascii.hexlify(x_raw), 16)
-    y = int(binascii.hexlify(y_raw), 16)
-    public_numbers = ec.EllipticCurvePublicNumbers(x, y, _lesc_curve())
-    return public_numbers.public_key(default_backend())
-
-
-def private_address_resolves(peer_addr, irk):
-    """
-    Checks if the given peer address can be resolved with the IRK
-
-    Private Resolvable Peer Addresses are in the format
-    [4x:xx:xx:yy:yy:yy], where 4x:xx:xx is a random number hashed with the IRK to generate yy:yy:yy
-    This function checks if the random number portion hashed with the IRK equals the hashed part of the address
-
-    :param peer_addr: The peer address to check
-    :param irk: The identity resolve key to try
-    :return: True if it resolves, False if not
-    """
-    # prand consists of the first 3 MSB bytes of the peer address
-    p_rand = str(bytearray(peer_addr.addr[:3]))
-    # the calculated hash is the last 3 LSB bytes of the peer address
-    addr_hash = str(bytearray(peer_addr.addr[3:]))
-    # IRK is stored in little-endian bytearray, convert to string and reverse
-    irk = str(irk)[::-1]
-    local_hash = ah(irk, p_rand)
-    return local_hash == addr_hash
 
 
 class SecurityParameters(object):
@@ -153,9 +73,9 @@ class SecurityManager(object):
         self.keyset = nrf_types.BLEGapSecKeyset()
         self.bond_db_entry = None
         self._security_level = SecurityLevel.NO_ACCESS
-        self._private_key = ec.generate_private_key(_lesc_curve, default_backend())
+        self._private_key = smp_crypto.lesc_generate_private_key()
         self._public_key = self._private_key.public_key()
-        self.keyset.own_keys.public_key.key = pubkey_to_raw(self._public_key)
+        self.keyset.own_keys.public_key.key = smp_crypto.lesc_pubkey_to_raw(self._public_key)
 
     """
     Events
@@ -314,7 +234,7 @@ class SecurityManager(object):
             if peer_address.addr_type in [nrf_types.BLEGapAddrTypes.random_static, nrf_types.BLEGapAddrTypes.public]:
                 if r.peer_addr == peer_address:
                     return r
-            elif private_address_resolves(peer_address, r.bonding_data.peer_id.irk):
+            elif smp_crypto.private_address_resolves(peer_address, r.bonding_data.peer_id.irk):
                 logger.info("Resolved Peer ID to {}".format(r.peer_addr))
                 return r
 
@@ -380,11 +300,9 @@ class SecurityManager(object):
         """
         :type event: nrf_events.GapEvtLescDhKeyRequest
         """
-        peer_public_key = pubkey_from_raw(event.remote_public_key.key)
-        shared_secret = self._private_key.exchange(ec.ECDH(), peer_public_key)
-        # Convert shared secret from big endian to little endian
-        shared_secret = shared_secret[::-1]
-        self.ble_device.ble_driver.ble_gap_lesc_dhkey_reply(event.conn_handle, nrf_types.BLEGapDhKey(shared_secret))
+        peer_public_key = smp_crypto.lesc_pubkey_from_raw(event.remote_public_key.key)
+        dh_key = smp_crypto.lesc_compute_dh_key(self._private_key, peer_public_key, little_endian=True)
+        self.ble_device.ble_driver.ble_gap_lesc_dhkey_reply(event.conn_handle, nrf_types.BLEGapDhKey(dh_key))
 
     def _on_conn_sec_status(self, driver, event):
         """
